+++ polars_utils/validation/actions.py
@@ -170,3 +170,25 @@ def make_col_has_fewer_nulls_validator(
         return nulls_col1 < nulls_col2
 
     return validate_col_has_fewer_nulls
+
+
+def make_convert_col_to_integers_preprocessor(
+    column: str,
+) -> Callable[[pl.DataFrame], pl.DataFrame]:
+    """
+    Creates a preprocessor function that converts a specified column to integers.
+
+    This is used to preprocess the year partition column for validation, allowing
+    for checks that require the column to be in a numerical format.
+
+    Args:
+        column (str): the name of the column to convert
+
+    Returns:
+        Callable[[pl.DataFrame], pl.DataFrame]: the inner function which converts the specified column to integers
+    """
+
+    def convert_col_to_integers(df: pl.DataFrame) -> pl.DataFrame:
+        return df.with_columns(pl.col(column).cast(pl.Int64))
+
+    return convert_col_to_integers
