+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/validate_04_estimates.py
@@ -1,6 +1,5 @@
 import sys
 from datetime import date, datetime
-from typing import Callable
 
 import pointblank as pb
 import polars as pl
@@ -325,7 +324,7 @@ def other_validation(
         )
         .col_vals_between(
             columns=PartitionKeys.year,
-            pre=make_convert_col_to_integers_preprocessor(PartitionKeys.year),
+            pre=vl.make_convert_col_to_integers_preprocessor(PartitionKeys.year),
             left=2013,
             right=int(datetime.now().year),
             brief="Year should be between 2013 and current year",
@@ -503,28 +502,6 @@ def estimates_percentage_expressions(
     return ((expr >= pcts[0]) & (expr <= pcts[1])).over(
         pl.col(IndCqcColumns.cqc_location_import_date)
     )
-
-
-def make_convert_col_to_integers_preprocessor(
-    column: str,
-) -> Callable[[pl.DataFrame], pl.DataFrame]:
-    """
-    Creates a preprocessor function that converts a specified column to integers.
-
-    This is used to preprocess the year partition column for validation, allowing
-    for checks that require the column to be in a numerical format.
-
-    Args:
-        column (str): the name of the column to convert
-
-    Returns:
-        Callable[[pl.DataFrame], pl.DataFrame]: the inner function which converts the specified column to integers
-    """
-
-    def convert_col_to_integers(df: pl.DataFrame) -> pl.DataFrame:
-        return df.with_columns(pl.col(column).cast(pl.Int64))
-
-    return convert_col_to_integers
 
 
 if __name__ == "__main__":
