+++ tests/test_polars_utils/test_validation_actions.py
@@ -4,9 +4,11 @@ from unittest.mock import ANY, patch
 import pointblank as pb
 import polars as pl
 import polars.testing as pl_testing
+import pytest
 
 import polars_utils.validation.actions as job
 from polars_utils.validation import actions as vl
+from utils.column_names.ind_cqc_pipeline_columns import PartitionKeys
 
 SRC_PATH = "polars_utils.validation.actions"
 
@@ -240,6 +242,20 @@ class TestMakeColHasFewerNullsValidator(TestValidate):
         result_fail = validator_fail(df_fail)
         # Then
         self.assertFalse(result_fail)
+
+
+class TestMakeConvertColToIntegersPreprocessor:
+    def test_make_convert_col_to_integers_preprocessor(self):
+        preprocessor = job.make_convert_col_to_integers_preprocessor(PartitionKeys.year)
+        df = pl.DataFrame({PartitionKeys.year: ["2020", "2021", "2022"]})
+        result = preprocessor(df)
+        expected = pl.DataFrame({PartitionKeys.year: [2020, 2021, 2022]})
+        pl_testing.assert_frame_equal(result, expected)
+
+    def test_convert_col_to_integers_with_non_numeric_values(self):
+        df = pl.DataFrame({PartitionKeys.year: ["2020", "invalid", "2022"]})
+        with pytest.raises(pl.exceptions.InvalidOperationError):
+            job.make_convert_col_to_integers_preprocessor(PartitionKeys.year)(df)
 
 
 if __name__ == "__main__":
